Standardize settings.
This is the version I have locally and have been using for months. I kept meaning to "finish" this and never submitted it, but it makes more sense to at least match between my local copy and the "real" version! All I did was rename the options to a safer format (with the full name of this package included so we don't accidentally overlap with some other YouTube package), and add explicit redirect_uri settings (I believe that requirement was added in httr2 after we worked on this).

Still to-do: Replace a lot of this with {nectar} (and/or use beekeeper to re-construct the package).